git pull hot & db schema update
