+++ db/container-content/config.sql
@@ -1,0 +1,1 @@
+CREATE EXTENSION tablefunc;
